Get rid of compiler warnings

+++ src/Db/MysqlConnection.cxx
@@ -1,4 +1,4 @@
-// $Header: /nfs/slac/g/glast/ground/cvs/rdbModel/src/Db/MysqlConnection.cxx,v 1.61 2008/04/21 20:41:58 jrb Exp $
+// $Header: /nfs/slac/g/glast/ground/cvs/GlastRelease-scons/rdbModel/src/Db/MysqlConnection.cxx,v 1.62 2008/12/04 20:23:10 jrb Exp $
 #ifdef  WIN32
 #include <windows.h>
 #endif
@@ -104,7 +104,8 @@ void addArg(bool literal, const std::string arg, std::string& sqlString) {
   std::string mysqlEscape(MYSQL* my, std::string const& s)
   {
     char * asciiz = new char[s.size()*2+1];
-    unsigned sz = mysql_real_escape_string(my, asciiz, s.c_str(), s.size());
+    unsigned sz = (unsigned) 
+      mysql_real_escape_string(my, asciiz, s.c_str(), s.size());
     std::string result(asciiz, sz);
     delete [] asciiz;
     return result;
@@ -1014,7 +1015,7 @@ namespace rdbModel {
         // compare sql size to enum value sizes
         sqlSize = extractSize(sqlType);
         for (unsigned i = 0; i < choices.size(); i++) {
-          if (sqlSize < choices[i].size() ) {
+          if ((unsigned) sqlSize < choices[i].size() ) {
             m_matchReturn = MATCHfail;
             (*m_err) << "sql type " << sqlType << " too small for this Enum "
                     << std::endl;
